requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.10
django-cors-headers==4.3.1
django-crontab==0.7.1
django-dotenv==1.4.2
django-kronos==1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
kronos==0.2.2
mysqlclient==2.2.4
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
